Further adaptation to Mac OS X: It seems that on Mac OS tooltips on menu items must not contain HTML code. Therefore, this has been removed where it was used.
Second, the method that creates the JMenuBar in BaseFrame has been split into several smaller methods, each of which being protected in order to ease the creation a more customized implementing class.

The menu item "Save" has been renamed to "Save as..." because this might be confusing as such an item cannot be predefined in a generic graphical user interface (it is not clear if all programs really open files and allow saving the results in the identical file again, so far I only know examples where some file is opened and some other file format or file is written). Furthermore, some users might think that their original work might get lost when clicking on "save".

git-svn-id: svn://rarepos.cs.uni-tuebingen.de/SysBio/trunk@676 e728d8a2-de87-4acf-9dc5-8dc47f4ff2f0