The default behavior of the method saveFile is now that it is redirected to saveFileAs in BaseFrame. It is no longer abstract. The reason is that not every program allows users to store some content in the identical file again that was parsed before. Hence, this method should be overriden if necessary, but not every program should have to implement it.
In SBML2GraphML I now followed the original approach by Clemens that a compartment is drawn if it is contained within some other compartment, i.e., the outer most compartment is not painted by default, except there is a compartment glyph for it. As long as we stick with default layouts and do not strictly follow what is defined by the SBML layout extension, this approach seems to be more meaningful to me. In the next version, however, we should only draw what is defined in the Layout extension.

git-svn-id: svn://rarepos.cs.uni-tuebingen.de/SysBio/trunk@972 e728d8a2-de87-4acf-9dc5-8dc47f4ff2f0